POC for backend connection for storage
This is functional, but will not work for more than 1 message at a time as it overloads the recipient. Instead of sending right to the game after you need to enqueue each message and send them once the back end has responded with a success or failure.

There is also an issue with sending a message after the first message. The connection is CLOSING or CLOSED which is not good.