Added Socket Function Triggering Code and Excessive Message Handler
Sicne you cannot send a message if there is one being sent you must wait. This was added in this patch, with a recursive call to itself.

I have also added in an event trigger so you can use promise style code while waiting for messages, so you know that your data has been sent to the game.